Use std math functions

+++ sr/Texture.hpp
@@ -245,8 +245,8 @@ namespace sr
 
                 if (sampler->filter == Sampler::Filter::Point)
                 {
-                    const uint32_t textureX = static_cast<uint32_t>(roundf(u));
-                    const uint32_t textureY = static_cast<uint32_t>(roundf(v));
+                    const uint32_t textureX = static_cast<uint32_t>(std::round(u));
+                    const uint32_t textureY = static_cast<uint32_t>(std::round(v));
                     return getPixel(textureX, textureY, 0);
                 }
                 else
@@ -358,12 +358,12 @@ namespace sr
                     for (uint32_t x = 0; x < dstWidth; ++x, pixel += 2, dst += 1)
                     {
                         float r = 0.0F;
-                        r += powf(pixel[0] / 255.0F, GAMMA);
-                        r += powf(pixel[1] / 255.0F, GAMMA);
-                        r += powf(pixel[pitch + 0] / 255.0F, GAMMA);
-                        r += powf(pixel[pitch + 1] / 255.0F, GAMMA);
+                        r += std::pow(pixel[0] / 255.0F, GAMMA);
+                        r += std::pow(pixel[1] / 255.0F, GAMMA);
+                        r += std::pow(pixel[pitch + 0] / 255.0F, GAMMA);
+                        r += std::pow(pixel[pitch + 1] / 255.0F, GAMMA);
                         r /= 4.0F;
-                        dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
+                        dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
                     }
                 }
             }
@@ -377,10 +377,10 @@ namespace sr
                     const uint8_t* pixel = src;
 
                     float r = 0.0F;
-                    r += powf(pixel[0] / 255.0F, GAMMA);
-                    r += powf(pixel[pitch + 0] / 255.0F, GAMMA);
+                    r += std::pow(pixel[0] / 255.0F, GAMMA);
+                    r += std::pow(pixel[pitch + 0] / 255.0F, GAMMA);
                     r /= 2.0F;
-                    dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
+                    dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
                 }
             }
             else if (dstWidth > 0)
@@ -389,10 +389,10 @@ namespace sr
                 for (uint32_t x = 0; x < dstWidth; ++x, pixel += 2, dst += 1)
                 {
                     float r = 0.0F;
-                    r += powf(pixel[0] / 255.0F, GAMMA);
-                    r += powf(pixel[1] / 255.0F, GAMMA);
+                    r += std::pow(pixel[0] / 255.0F, GAMMA);
+                    r += std::pow(pixel[1] / 255.0F, GAMMA);
                     r /= 2.0F;
-                    dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
+                    dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
                 }
             }
         }
@@ -419,36 +419,36 @@ namespace sr
 
                         if (pixel[3] > 0)
                         {
-                            r += powf(pixel[0] / 255.0F, GAMMA);
-                            g += powf(pixel[1] / 255.0F, GAMMA);
-                            b += powf(pixel[2] / 255.0F, GAMMA);
+                            r += std::pow(pixel[0] / 255.0F, GAMMA);
+                            g += std::pow(pixel[1] / 255.0F, GAMMA);
+                            b += std::pow(pixel[2] / 255.0F, GAMMA);
                             pixels += 1.0F;
                         }
                         a += pixel[3];
 
                         if (pixel[7] > 0)
                         {
-                            r += powf(pixel[4] / 255.0F, GAMMA);
-                            g += powf(pixel[5] / 255.0F, GAMMA);
-                            b += powf(pixel[6] / 255.0F, GAMMA);
+                            r += std::pow(pixel[4] / 255.0F, GAMMA);
+                            g += std::pow(pixel[5] / 255.0F, GAMMA);
+                            b += std::pow(pixel[6] / 255.0F, GAMMA);
                             pixels += 1.0F;
                         }
                         a += pixel[7];
 
                         if (pixel[pitch + 3] > 0)
                         {
-                            r += powf(pixel[pitch + 0] / 255.0F, GAMMA);
-                            g += powf(pixel[pitch + 1] / 255.0F, GAMMA);
-                            b += powf(pixel[pitch + 2] / 255.0F, GAMMA);
+                            r += std::pow(pixel[pitch + 0] / 255.0F, GAMMA);
+                            g += std::pow(pixel[pitch + 1] / 255.0F, GAMMA);
+                            b += std::pow(pixel[pitch + 2] / 255.0F, GAMMA);
                             pixels += 1.0F;
                         }
                         a += pixel[pitch + 3];
 
                         if (pixel[pitch + 7] > 0)
                         {
-                            r += powf(pixel[pitch + 4] / 255.0F, GAMMA);
-                            g += powf(pixel[pitch + 5] / 255.0F, GAMMA);
-                            b += powf(pixel[pitch + 6] / 255.0F, GAMMA);
+                            r += std::pow(pixel[pitch + 4] / 255.0F, GAMMA);
+                            g += std::pow(pixel[pitch + 5] / 255.0F, GAMMA);
+                            b += std::pow(pixel[pitch + 6] / 255.0F, GAMMA);
                             pixels += 1.0F;
                         }
                         a += pixel[pitch + 7];
@@ -459,9 +459,9 @@ namespace sr
                             g /= pixels;
                             b /= pixels;
                             a *= 0.25F;
-                            dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
-                            dst[1] = static_cast<uint8_t>(roundf(powf(g, 1.0F / GAMMA) * 255.0F));
-                            dst[2] = static_cast<uint8_t>(roundf(powf(b, 1.0F / GAMMA) * 255.0F));
+                            dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
+                            dst[1] = static_cast<uint8_t>(std::round(std::pow(g, 1.0F / GAMMA) * 255.0F));
+                            dst[2] = static_cast<uint8_t>(std::round(std::pow(b, 1.0F / GAMMA) * 255.0F));
                             dst[3] = static_cast<uint8_t>(a);
                         }
                         else
@@ -491,18 +491,18 @@ namespace sr
 
                     if (pixel[3] > 0)
                     {
-                        r += powf(pixel[0] / 255.0F, GAMMA);
-                        g += powf(pixel[1] / 255.0F, GAMMA);
-                        b += powf(pixel[2] / 255.0F, GAMMA);
+                        r += std::pow(pixel[0] / 255.0F, GAMMA);
+                        g += std::pow(pixel[1] / 255.0F, GAMMA);
+                        b += std::pow(pixel[2] / 255.0F, GAMMA);
                         pixels += 1.0F;
                     }
                     a = pixel[3];
 
                     if (pixel[pitch + 3] > 0)
                     {
-                        r += powf(pixel[pitch + 0] / 255.0F, GAMMA);
-                        g += powf(pixel[pitch + 1] / 255.0F, GAMMA);
-                        b += powf(pixel[pitch + 2] / 255.0F, GAMMA);
+                        r += std::pow(pixel[pitch + 0] / 255.0F, GAMMA);
+                        g += std::pow(pixel[pitch + 1] / 255.0F, GAMMA);
+                        b += std::pow(pixel[pitch + 2] / 255.0F, GAMMA);
                         pixels += 1.0F;
                     }
                     a += pixel[pitch + 3];
@@ -513,9 +513,9 @@ namespace sr
                         g /= pixels;
                         b /= pixels;
                         a *= 0.5F;
-                        dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
-                        dst[1] = static_cast<uint8_t>(roundf(powf(g, 1.0F / GAMMA) * 255.0F));
-                        dst[2] = static_cast<uint8_t>(roundf(powf(b, 1.0F / GAMMA) * 255.0F));
+                        dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
+                        dst[1] = static_cast<uint8_t>(std::round(std::pow(g, 1.0F / GAMMA) * 255.0F));
+                        dst[2] = static_cast<uint8_t>(std::round(std::pow(b, 1.0F / GAMMA) * 255.0F));
                         dst[3] = static_cast<uint8_t>(a);
                     }
                     else
@@ -540,18 +540,18 @@ namespace sr
 
                     if (pixel[3] > 0)
                     {
-                        r += powf(pixel[0] / 255.0F, GAMMA);
-                        g += powf(pixel[1] / 255.0F, GAMMA);
-                        b += powf(pixel[2] / 255.0F, GAMMA);
+                        r += std::pow(pixel[0] / 255.0F, GAMMA);
+                        g += std::pow(pixel[1] / 255.0F, GAMMA);
+                        b += std::pow(pixel[2] / 255.0F, GAMMA);
                         pixels += 1.0F;
                     }
                     a += pixel[3];
 
                     if (pixel[7] > 0)
                     {
-                        r += powf(pixel[4] / 255.0F, GAMMA);
-                        g += powf(pixel[5] / 255.0F, GAMMA);
-                        b += powf(pixel[6] / 255.0F, GAMMA);
+                        r += std::pow(pixel[4] / 255.0F, GAMMA);
+                        g += std::pow(pixel[5] / 255.0F, GAMMA);
+                        b += std::pow(pixel[6] / 255.0F, GAMMA);
                         pixels += 1.0F;
                     }
                     a += pixel[7];
@@ -562,9 +562,9 @@ namespace sr
                         g /= pixels;
                         b /= pixels;
                         a *= 0.5F;
-                        dst[0] = static_cast<uint8_t>(roundf(powf(r, 1.0F / GAMMA) * 255.0F));
-                        dst[1] = static_cast<uint8_t>(roundf(powf(g, 1.0F / GAMMA) * 255.0F));
-                        dst[2] = static_cast<uint8_t>(roundf(powf(b, 1.0F / GAMMA) * 255.0F));
+                        dst[0] = static_cast<uint8_t>(std::round(std::pow(r, 1.0F / GAMMA) * 255.0F));
+                        dst[1] = static_cast<uint8_t>(std::round(std::pow(g, 1.0F / GAMMA) * 255.0F));
+                        dst[2] = static_cast<uint8_t>(std::round(std::pow(b, 1.0F / GAMMA) * 255.0F));
                         dst[3] = static_cast<uint8_t>(a);
                     }
                     else
